requirements.txt
numpy
tifffile
pyzmq==19.0.2